requirements.txt
# Core dependencies for SpikeFormer Neuromorphic Kit
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0
pyyaml>=6.0
click>=8.0.0
rich>=12.0.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0
wandb>=0.15.0
mlflow>=2.0.0
prometheus-client>=0.14.0
psutil>=5.8.0